move KNNGraph class to Space module
